Refactors service routing and API calls on the front-end.
